More refactoring and updated the API to allow 64 bit indexes and added more unit tests

+++ pkg/csstack/csstack_test.go
@@ -464,7 +464,7 @@ func TestCSStackFindIndices(t *testing.T) {
 func TestNewCSStackFromSlice(t *testing.T) {
 	items := []int{1, 2, 3, 4, 5}
 	cs := csstack.NewCSStackFromSlice(items)
-	if cs.Size() != len(items) {
+	if cs.Size() != uint64(len(items)) {
 		t.Fatalf(errExpectedSizeX, len(items), cs.Size())
 	}
 	slice := cs.ToSlice()
